Added tools and removed duplicate code
