+++ RealmModelGenerator/SwiftContentGenerator.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import AddressBook
 
 class SwiftContentGenerator {
     static let TAG = NSStringFromClass(SwiftContentGenerator)
@@ -19,7 +18,7 @@ class SwiftContentGenerator {
         self.entity = entity
     }
     
-    func getContent()->String
+    func getContent() -> String
     {
         appendHeader()
         appendAttributesAndAnnotations()
@@ -32,41 +31,12 @@ class SwiftContentGenerator {
     //MARK: Append header
     func appendHeader()
     {
-        appendHeaderComments()
+        content += Tools.getHeaderComments(entity, fileExtension: "swift")
         
         // Static import
         content += "import RealmSwift\n\n"
         content += "class " + entity.name + " : Object {\n"
         content += "  static let TAG = NSStringFromClass(" + entity.name + ");\n\n"
-    }
-    
-    func appendHeaderComments() {
-        content += "/**\n"
-        content += " *\t" + entity.name + ".swift\n"
-        content += " *\tModel version: " + entity.model.version + "\n"
-        
-        if let me = ABAddressBook.sharedAddressBook()?.me(){
-            
-            if let firstName = me.valueForProperty(kABFirstNameProperty as String) as? String{
-                content += " *\tCreate by \(firstName)"
-                if let lastName = me.valueForProperty(kABLastNameProperty as String) as? String{
-                    content += " \(lastName)"
-                }
-            }
-            
-            content += " on \(getTodayFormattedDay())\n *\tCopyright © \(getYear())"
-            
-            if let organization = me.valueForProperty(kABOrganizationProperty as String) as? String{
-                content += " \(organization)"
-            } else {
-                content += " QuarkWorks"
-            }
-            
-            content += ". All rights reserved.\n"
-        }
-        
-        content += " *\tModel file Generated using Realm Model Genrator.\n"
-        content += " */\n\n"
     }
     
     //MARK: Append attributes and relicationships
@@ -160,18 +130,5 @@ class SwiftContentGenerator {
         content += "  override static func ignoredProperties() -> [String] {\n"
         content += "    return [" + ignoredProperties + "]\n"
         content += "  }\n\n"
-    }
-    
-    // Returns the current year as String
-    func getYear() -> String
-    {
-        return "\(NSCalendar.currentCalendar().component(.Year, fromDate: NSDate()))"
-    }
-    
-    // Returns today date in the format dd/mm/yyyy
-    func getTodayFormattedDay() -> String
-    {
-        let components = NSCalendar.currentCalendar().components([.Day, .Month, .Year], fromDate: NSDate())
-        return "\(components.day)/\(components.month)/\(components.year)"
     }
 }
